requirements.txt
python-binance==1.0.19
requests==2.31.0
click==8.1.7
python-dotenv==1.0.0
rich==13.7.0